chore: sort the example for clarity
